fix duplicated code by moving it to SimpleActivity.kt
